Adapt           -removed set_leaf()           -added DESTROY debugging print Refresh           -added warning if block should have been deleted           -removed call to check_new_() Output           -BUG FIX time_next was int not double in ScheduleInterval Mesh           -Index consolidated fprintf's to improve performance Parameters           -initialize ghosts to 0 if Field:ghosts not defined

+++ src/Cello/io_ScheduleInterval.cpp
@@ -103,7 +103,7 @@ double ScheduleInterval::update_timestep ( double time, double dt) const throw()
 
       if (in_range) {
 
-	int time_next = this->time_next();
+	double time_next = this->time_next();
 
 	if (time < time_next && time_next < time + dt) {
 	  new_dt = time_next - time;
